moved initializeJqueryCvuRadio
