+++ app/javascript/controllers/cypress_controller.js
@@ -8,7 +8,6 @@ export default class extends Controller {
     if (this.initializedValue) return
     this.initializedValue = true
 
-    cypress.initializeJqueryCvuRadio?.()
     cypress.initializeProductTable?.()
     cypress.reticulateSplines?.()
     cypress.initializeMeasureSelection?.()
